update Poisson compiles
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/LaplaceVector_merge@3399 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/application_cpp/NativeCppHeatEquationSolverMPI.cpp
@@ -324,7 +324,7 @@ void testHeatEquation(size_t dim, size_t start_level, size_t end_level, double b
       myHESolver->initScreen();
 
       // Construct a grid
-      myHESolver->constructGrid(*myBoundingBox, l);
+      myHESolver->constructGrid(*myBoundingBox, (int)l);
 
       // init the basis functions' coefficient vector (start solution)
       alpha = new sg::base::DataVector(myHESolver->getNumberGridPoints());
@@ -479,7 +479,7 @@ void testPoissonEquation(size_t dim, size_t start_level, size_t end_level, doubl
       myPoisSolver->initScreen();
 
       // Construct a grid
-      myPoisSolver->constructGrid(*myBoundingBox, l);
+      myPoisSolver->constructGrid(*myBoundingBox, (int)l);
 
       // init the basis functions' coefficient vector (start solution)
       alpha = new sg::base::DataVector(myPoisSolver->getNumberGridPoints());
@@ -617,7 +617,7 @@ void testPoissonEquationAdapt(size_t dim, size_t start_level, std::string refine
     myPoisSolver->initScreen();
 
     // Construct a grid
-    myPoisSolver->constructGrid(*myBoundingBox, start_level);
+    myPoisSolver->constructGrid(*myBoundingBox, (int)start_level);
 
     // init the basis functions' coefficient vector (start solution)
     // init the basis functions' coefficient vector (start solution)
@@ -634,7 +634,7 @@ void testPoissonEquationAdapt(size_t dim, size_t start_level, std::string refine
         if (refine == "classic") {
           myPoisSolver->refineInitialGridSurplus(*alpha, -1, refine_thres);
         } else {
-          myPoisSolver->refineInitialGridSurplusToMaxLevel(*alpha, refine_thres, max_ref_level);
+          myPoisSolver->refineInitialGridSurplusToMaxLevel(*alpha, refine_thres, (int)max_ref_level);
         }
 
         std::cout << "Refined Grid size: " << myPoisSolver->getNumberGridPoints() << std::endl;
@@ -655,7 +655,7 @@ void testPoissonEquationAdapt(size_t dim, size_t start_level, std::string refine
         if (refine == "classic") {
           myPoisSolver->refineInitialGridSurplus(*alpha, -1, refine_thres);
         } else {
-          myPoisSolver->refineInitialGridSurplusToMaxLevel(*alpha, refine_thres, max_ref_level);
+          myPoisSolver->refineInitialGridSurplusToMaxLevel(*alpha, refine_thres, (int)max_ref_level);
         }
 
         std::cout << "Refined Grid size: " << myPoisSolver->getNumberGridPoints() << std::endl;
